chore: 프로젝트 초기 세팅
