+++ NewsFit/NewsFit/Resources/Assets.xcassets/AppIcon.appiconset/Contents.json
