+++ NewsFit/NewsFit/Resources/Assets.xcassets/Contents.json
